fix(): move func
